change some details
I found that the name in the table is divided into last name and first name, so I wrote the names separately. I'll try to put the three scores together in parentheses, otherwise writing them separately would look a bit messy. After that, I'll use the try catch method to import the data, and I'll see if my approach is correct.