menu
This class implements a simple text-based menu system for interacting with student records. It allows users to read student data from a file, display all students, filter students by total marks, display top and bottom 5 students based on their marks, and exit the program. I also added a detail that if the threshold input number is not within 0-100, it is invalid and needs to be re entered.