the blank part
There are three students whose grades are blank in the file. When I read the data, the blank line will not be displayed in the overall grade, so I added some conditions to change the blank part to zero so that it will be counted as zero in the overall grade. And i add some references and detials